backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
djangorestframework==3.12.1
idna==2.10
joblib==0.17.0
numpy==1.19.2
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.25.10
django-cors-headers==3.6.0
gunicorn==20.0.4
psycopg2-binary==2.8.6